Backport recent CI fixes (#1758): #1718 #1731 #1739 #1742
* RTD: Fix GA Integration (#1718)

GA was dropped from RTD in early Oct, 2024. This adds it again.

* CI: Move Ubuntu 20.04 workflows to 22.04 (#1731)

* Reactivate MPICH runner (gcc13_py312_mpich_h5_ad2) (#1739)

* Revert "CI: Disable MPICH for now"

This reverts commit de9e7ad02944e91206bea3a603b4d0ea10771d21.

* We shall try installin MPICH from Spack for now

Ubuntu package broken https://bugs.launchpad.net/ubuntu/+source/mpich/+bug/2072338

* Take out version checks

* Revert "Take out version checks"

This reverts commit b22e4b32970975992ec888dd65f01a9ba09b13c6.

* Revert "We shall try installin MPICH from Spack for now"

This reverts commit 33b847fe7a5f6e81db76eb4fe4c2ec5c156c4001.

* Alternative approach: self-compiled MPICH

* Move MPICH installation to separate file

* Fix broken CI: Support for old CMake versions (#1742)

* Check for broken CI

* Deactivate dependencies that are not yet compatible with CMake 4.0

* Fixes

* Upgrade Spack

* Fix warning

* Fixes after upgrading spack

---------

+++ examples/5_write_parallel.cpp
@@ -43,7 +43,7 @@ int main(int argc, char *argv[])
     // global data set to write: [MPI_Size * 10, 300]
     // each rank writes a 10x300 slice with its MPI rank as values
     auto const value = float(mpi_size);
-    std::vector<float> local_data(10 * 300, value);
+    std::vector<float> local_data(size_t(10) * 300, value);
     if (0 == mpi_rank)
         cout << "Set up a 2D array with 10x300 elements per MPI rank ("
              << mpi_size << "x) that will be written to disk\n";
